Rename files for cl vs aoa.
cl_aoa_lab.m is used to plot the cl obtained with all the methods used,
and compare them with cl obtained from lab data, for 16 m/s and 25 m/s.

cl_aoa_re.m is used to plot the cl obtained with all the methods used,
for different values of Reynolds number.