Fix computations of cl and cd from cp.
Calculations of axial and normal forces from the pressure distribution
were wrong, leading to uncorrect values of cd and cl.

Now the cd converges to 0 when the number of panels used increase, as it
should. The convergence study is now focus on cd only, and plot it on a
semilogy axis.